add BaseLearner() so that other meta learners inherit from it
